[dev-launcher] move session fns from dev-menu to dev-launcher
