+++ packages/expo-dev-launcher/bundle/native-modules/__mocks__/DevMenuInternal.ts
@@ -1,5 +1,2 @@
-export const setSessionAsync = jest.fn().mockResolvedValue(null);
-export const restoreSessionAsync = jest.fn().mockResolvedValue(null);
-export const getAuthSchemeAsync = jest.fn().mockResolvedValue('123');
 export const getSettingsAsync = jest.fn().mockResolvedValue({});
 export const setSettingsAsync = jest.fn().mockResolvedValue({});
