--- packages/expo-dev-launcher/bundle/native-modules/__mocks__/DevMenuWebBrowser.ts
@@ -1,1 +1,0 @@
-export const openAuthSessionAsync = jest.fn().mockResolvedValue('123');
