+++ packages/expo-dev-launcher/ios/ReactNative/EXDevLauncherRCTBridge.m
@@ -19,7 +19,7 @@
 - (RCTDevSettings *)devSettings
 {
   //  uncomment below to enable fast refresh for development builds of DevLauncher
-  // return super.devSettings;
+   return super.devSettings;
  return nil;
 }
