+++ packages/expo-dev-launcher/setupTests.ts
@@ -31,7 +31,6 @@ jest.mock('react-native/Libraries/Components/Switch/Switch', () => {
 
 jest.mock('./bundle/native-modules/DevLauncherInternal');
 jest.mock('./bundle/native-modules/DevMenuInternal');
-jest.mock('./bundle/native-modules/DevMenuWebBrowser');
 jest.mock('./bundle/native-modules/DevMenu');
 
 const MOCK_INITIAL_METRICS = {
