+++ packages/expo-dev-menu/ios/Modules/DevMenuInternalModule.swift
@@ -3,13 +3,6 @@ import SafariServices
 
 @objc(DevMenuInternalModule)
 public class DevMenuInternalModule: NSObject, RCTBridgeModule {
-  @objc
-  var redirectResolve: RCTPromiseResolveBlock?
-  @objc
-  var redirectReject: RCTPromiseRejectBlock?
-  @objc
-  var authSession: SFAuthenticationSession?
-
   public static func moduleName() -> String! {
     return "ExpoDevMenuInternal"
   }
@@ -18,12 +11,6 @@ public class DevMenuInternalModule: NSObject, RCTBridgeModule {
   public static func requiresMainQueueSetup() -> Bool {
     return true
   }
-
-  private static var fontsWereLoaded = false
-  private static let sessionKey = "expo-dev-menu.session"
-  private static let userLoginEvent = "expo.dev-menu.user-login"
-  private static let userLogoutEvent = "expo.dev-menu.user-logout"
-  private static let defaultScheme = "expo-dev-menu"
 
   let manager: DevMenuManager
 
@@ -45,44 +32,6 @@ public class DevMenuInternalModule: NSObject, RCTBridgeModule {
     let doesDeviceSupportKeyCommands = false
 #endif
     return ["doesDeviceSupportKeyCommands": doesDeviceSupportKeyCommands]
-  }
-
-  @objc
-  func loadFontsAsync(_ resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
-    if DevMenuInternalModule.fontsWereLoaded {
-      resolve(nil)
-      return
-    }
-
-    let fonts = ["MaterialCommunityIcons", "Ionicons"]
-    for font in fonts {
-      guard let path = DevMenuUtils.resourcesBundle()?.path(forResource: font, ofType: "ttf") else {
-        reject("ERR_DEVMENU_CANNOT_FIND_FONT", "Font file for '\(font)' doesn't exist.", nil)
-        return
-      }
-      guard let data = FileManager.default.contents(atPath: path) else {
-        reject("ERR_DEVMENU_CANNOT_OPEN_FONT_FILE", "Could not open '\(path)'.", nil)
-        return
-      }
-
-      guard let provider = CGDataProvider(data: data as CFData) else {
-        reject("ERR_DEVMENU_CANNOT_CREATE_FONT_PROVIDER", "Could not create font provider for '\(font)'.", nil)
-        return
-      }
-      guard let cgFont = CGFont(provider) else {
-        reject("ERR_DEVMENU_CANNOT_CREATE_FONT", "Could not create font for '\(font)'.", nil)
-        return
-      }
-
-      var error: Unmanaged<CFError>?
-      if !CTFontManagerRegisterGraphicsFont(cgFont, &error) {
-        reject("ERR_DEVMENU_CANNOT_ADD_FONT", "Could not create font from loaded data for '\(font)'. '\(error.debugDescription)'.", nil)
-        return
-      }
-    }
-
-    DevMenuInternalModule.fontsWereLoaded = true
-    resolve(nil)
   }
 
   @objc
@@ -160,49 +109,6 @@ public class DevMenuInternalModule: NSObject, RCTBridgeModule {
     resolve(nil)
   }
 
-  @objc
-  func setSessionAsync(_ session: [String: Any]?, resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
-    do {
-      try manager.expoSessionDelegate.setSessionAsync(session)
-      resolve(nil)
-    } catch let error {
-      reject("ERR_DEVMENU_CANNOT_SAVE_SESSION", error.localizedDescription, error)
-    }
-  }
-
-  @objc
-  func restoreSessionAsync(_ resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
-    resolve(manager.expoSessionDelegate.restoreSession())
-  }
-
-  @objc
-  func getAuthSchemeAsync(_ resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
-    guard let urlTypesArray = Bundle.main.infoDictionary?["CFBundleURLTypes"] as? [NSDictionary] else {
-      resolve(DevMenuInternalModule.defaultScheme)
-      return
-    }
-
-    if (urlTypesArray
-          .contains(where: { ($0["CFBundleURLSchemes"] as? [String] ?? [])
-                      .contains(DevMenuInternalModule.defaultScheme) })) {
-      resolve(DevMenuInternalModule.defaultScheme)
-      return
-    }
-
-    for urlType in urlTypesArray {
-      guard let schemes = urlType["CFBundleURLSchemes"] as? [String] else {
-        continue
-      }
-
-      if schemes.first != nil {
-        resolve(schemes.first)
-        return
-      }
-    }
-
-    resolve(DevMenuInternalModule.defaultScheme)
-  }
-  
   @objc
   func getBuildInfoAsync(_ resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
     let bridge = manager.delegate?.appBridge?(forDevMenuManager: self.manager)
